refactor: use one template for HA/non-HA VPN envoy
