+++ pkg/component/networking/vpn/envoy/templates/envoy.yaml.tpl
