+++ pkg/component/shared/kubeapiserver_test.go
@@ -1156,6 +1156,7 @@ authorizers:
 			nodeNetworkCIDRs               []net.IPNet
 			podNetworkCIDRs                []net.IPNet
 			serviceNetworkCIDRs            []net.IPNet
+			seedPodNetwork                 string
 			etcdEncryptionKeyRotationPhase gardencorev1beta1.CredentialsRotationPhase
 			wantScaleDown                  bool
 		)
@@ -1173,6 +1174,7 @@ authorizers:
 			nodeNetworkCIDRs = []net.IPNet{{IP: net.ParseIP("10.250.0.0"), Mask: net.CIDRMask(24, 32)}}
 			serviceNetworkCIDRs = []net.IPNet{{IP: net.ParseIP("10.0.2.0"), Mask: net.CIDRMask(24, 32)}}
 			podNetworkCIDRs = []net.IPNet{{IP: net.ParseIP("10.0.1.0"), Mask: net.CIDRMask(24, 32)}}
+			seedPodNetwork = "100.64.0.0/12"
 			wantScaleDown = false
 		})
 
@@ -1203,11 +1205,12 @@ authorizers:
 				kubeAPIServer.EXPECT().SetNodeNetworkCIDRs(gomock.Any())
 				kubeAPIServer.EXPECT().SetServiceNetworkCIDRs(gomock.Any())
 				kubeAPIServer.EXPECT().SetPodNetworkCIDRs(gomock.Any())
+				kubeAPIServer.EXPECT().SetSeedPodNetwork(gomock.Any())
 				kubeAPIServer.EXPECT().SetServerCertificateConfig(gomock.Any())
 				kubeAPIServer.EXPECT().SetServiceAccountConfig(gomock.Any())
 				kubeAPIServer.EXPECT().Deploy(ctx)
 
-				Expect(DeployKubeAPIServer(ctx, runtimeClient, namespace, kubeAPIServer, serviceAccountConfig, serverCertificateConfig, sniConfig, externalHostname, nodeNetworkCIDRs, serviceNetworkCIDRs, podNetworkCIDRs, nil, nil, etcdEncryptionKeyRotationPhase, wantScaleDown)).To(Succeed())
+				Expect(DeployKubeAPIServer(ctx, runtimeClient, namespace, kubeAPIServer, serviceAccountConfig, serverCertificateConfig, sniConfig, externalHostname, nodeNetworkCIDRs, serviceNetworkCIDRs, podNetworkCIDRs, seedPodNetwork, nil, nil, etcdEncryptionKeyRotationPhase, wantScaleDown)).To(Succeed())
 			},
 
 			Entry("nothing is set when deployment is not found",
@@ -1277,11 +1280,12 @@ authorizers:
 				kubeAPIServer.EXPECT().SetNodeNetworkCIDRs(gomock.Any())
 				kubeAPIServer.EXPECT().SetServiceNetworkCIDRs(gomock.Any())
 				kubeAPIServer.EXPECT().SetPodNetworkCIDRs(gomock.Any())
+				kubeAPIServer.EXPECT().SetSeedPodNetwork(gomock.Any())
 				kubeAPIServer.EXPECT().SetServerCertificateConfig(gomock.Any())
 				kubeAPIServer.EXPECT().SetServiceAccountConfig(gomock.Any())
 				kubeAPIServer.EXPECT().Deploy(ctx)
 
-				Expect(DeployKubeAPIServer(ctx, runtimeClient, namespace, kubeAPIServer, serviceAccountConfig, serverCertificateConfig, sniConfig, externalHostname, nodeNetworkCIDRs, serviceNetworkCIDRs, podNetworkCIDRs, nil, nil, etcdEncryptionKeyRotationPhase, wantScaleDown)).To(Succeed())
+				Expect(DeployKubeAPIServer(ctx, runtimeClient, namespace, kubeAPIServer, serviceAccountConfig, serverCertificateConfig, sniConfig, externalHostname, nodeNetworkCIDRs, serviceNetworkCIDRs, podNetworkCIDRs, seedPodNetwork, nil, nil, etcdEncryptionKeyRotationPhase, wantScaleDown)).To(Succeed())
 			},
 
 			Entry("no change due to already set",
@@ -1352,11 +1356,12 @@ authorizers:
 				kubeAPIServer.EXPECT().SetNodeNetworkCIDRs(gomock.Any())
 				kubeAPIServer.EXPECT().SetServiceNetworkCIDRs(gomock.Any())
 				kubeAPIServer.EXPECT().SetPodNetworkCIDRs(gomock.Any())
+				kubeAPIServer.EXPECT().SetSeedPodNetwork(gomock.Any())
 				kubeAPIServer.EXPECT().SetServerCertificateConfig(gomock.Any())
 				kubeAPIServer.EXPECT().SetServiceAccountConfig(gomock.Any())
 				kubeAPIServer.EXPECT().Deploy(ctx)
 
-				Expect(DeployKubeAPIServer(ctx, runtimeClient, namespace, kubeAPIServer, serviceAccountConfig, serverCertificateConfig, sniConfig, externalHostname, nodeNetworkCIDRs, serviceNetworkCIDRs, podNetworkCIDRs, nil, nil, etcdEncryptionKeyRotationPhase, wantScaleDown)).To(Succeed())
+				Expect(DeployKubeAPIServer(ctx, runtimeClient, namespace, kubeAPIServer, serviceAccountConfig, serverCertificateConfig, sniConfig, externalHostname, nodeNetworkCIDRs, serviceNetworkCIDRs, podNetworkCIDRs, seedPodNetwork, nil, nil, etcdEncryptionKeyRotationPhase, wantScaleDown)).To(Succeed())
 
 				if finalizeTest != nil {
 					finalizeTest()
@@ -1465,11 +1470,12 @@ authorizers:
 				kubeAPIServer.EXPECT().SetNodeNetworkCIDRs(gomock.Any())
 				kubeAPIServer.EXPECT().SetServiceNetworkCIDRs(gomock.Any())
 				kubeAPIServer.EXPECT().SetPodNetworkCIDRs(gomock.Any())
+				kubeAPIServer.EXPECT().SetSeedPodNetwork(gomock.Any())
 				kubeAPIServer.EXPECT().SetServerCertificateConfig(gomock.Any())
 				kubeAPIServer.EXPECT().SetServiceAccountConfig(gomock.Any())
 				kubeAPIServer.EXPECT().Deploy(ctx)
 
-				Expect(DeployKubeAPIServer(ctx, runtimeClient, namespace, kubeAPIServer, serviceAccountConfig, serverCertificateConfig, sniConfig, externalHostname, nodeNetworkCIDRs, serviceNetworkCIDRs, podNetworkCIDRs, nil, nil, etcdEncryptionKeyRotationPhase, wantScaleDown)).To(Succeed())
+				Expect(DeployKubeAPIServer(ctx, runtimeClient, namespace, kubeAPIServer, serviceAccountConfig, serverCertificateConfig, sniConfig, externalHostname, nodeNetworkCIDRs, serviceNetworkCIDRs, podNetworkCIDRs, seedPodNetwork, nil, nil, etcdEncryptionKeyRotationPhase, wantScaleDown)).To(Succeed())
 			})
 
 			It("It should deploy KubeAPIServer with the default resources appended to the passed resources", func() {
@@ -1494,6 +1500,7 @@ authorizers:
 				kubeAPIServer.EXPECT().SetNodeNetworkCIDRs(gomock.Any())
 				kubeAPIServer.EXPECT().SetServiceNetworkCIDRs(gomock.Any())
 				kubeAPIServer.EXPECT().SetPodNetworkCIDRs(gomock.Any())
+				kubeAPIServer.EXPECT().SetSeedPodNetwork(gomock.Any())
 				kubeAPIServer.EXPECT().SetServerCertificateConfig(gomock.Any())
 				kubeAPIServer.EXPECT().SetServiceAccountConfig(gomock.Any())
 				kubeAPIServer.EXPECT().Deploy(ctx)
@@ -1507,7 +1514,7 @@ authorizers:
 					"deployments.apps",
 				}
 
-				Expect(DeployKubeAPIServer(ctx, runtimeClient, namespace, kubeAPIServer, serviceAccountConfig, serverCertificateConfig, sniConfig, externalHostname, nodeNetworkCIDRs, serviceNetworkCIDRs, podNetworkCIDRs, resourcesToEncrypt, encryptedResources, etcdEncryptionKeyRotationPhase, wantScaleDown)).To(Succeed())
+				Expect(DeployKubeAPIServer(ctx, runtimeClient, namespace, kubeAPIServer, serviceAccountConfig, serverCertificateConfig, sniConfig, externalHostname, nodeNetworkCIDRs, serviceNetworkCIDRs, podNetworkCIDRs, seedPodNetwork, resourcesToEncrypt, encryptedResources, etcdEncryptionKeyRotationPhase, wantScaleDown)).To(Succeed())
 			})
 		})
 
@@ -1521,11 +1528,12 @@ authorizers:
 				kubeAPIServer.EXPECT().SetNodeNetworkCIDRs(gomock.Any())
 				kubeAPIServer.EXPECT().SetServiceNetworkCIDRs(gomock.Any())
 				kubeAPIServer.EXPECT().SetPodNetworkCIDRs(gomock.Any())
+				kubeAPIServer.EXPECT().SetSeedPodNetwork(gomock.Any())
 				kubeAPIServer.EXPECT().SetServerCertificateConfig(gomock.Any())
 				kubeAPIServer.EXPECT().SetServiceAccountConfig(gomock.Any())
 				kubeAPIServer.EXPECT().Deploy(ctx)
 
-				Expect(DeployKubeAPIServer(ctx, runtimeClient, namespace, kubeAPIServer, serviceAccountConfig, serverCertificateConfig, sniConfig, externalHostname, nodeNetworkCIDRs, serviceNetworkCIDRs, podNetworkCIDRs, nil, nil, etcdEncryptionKeyRotationPhase, wantScaleDown)).To(Succeed())
+				Expect(DeployKubeAPIServer(ctx, runtimeClient, namespace, kubeAPIServer, serviceAccountConfig, serverCertificateConfig, sniConfig, externalHostname, nodeNetworkCIDRs, serviceNetworkCIDRs, podNetworkCIDRs, seedPodNetwork, nil, nil, etcdEncryptionKeyRotationPhase, wantScaleDown)).To(Succeed())
 			})
 		})
 
@@ -1539,11 +1547,12 @@ authorizers:
 				kubeAPIServer.EXPECT().SetNodeNetworkCIDRs(gomock.Any())
 				kubeAPIServer.EXPECT().SetServiceNetworkCIDRs(gomock.Any())
 				kubeAPIServer.EXPECT().SetPodNetworkCIDRs(gomock.Any())
+				kubeAPIServer.EXPECT().SetSeedPodNetwork(gomock.Any())
 				kubeAPIServer.EXPECT().SetServerCertificateConfig(serverCertificateConfig)
 				kubeAPIServer.EXPECT().SetServiceAccountConfig(gomock.Any())
 				kubeAPIServer.EXPECT().Deploy(ctx)
 
-				Expect(DeployKubeAPIServer(ctx, runtimeClient, namespace, kubeAPIServer, serviceAccountConfig, serverCertificateConfig, sniConfig, externalHostname, nodeNetworkCIDRs, serviceNetworkCIDRs, podNetworkCIDRs, nil, nil, etcdEncryptionKeyRotationPhase, wantScaleDown)).To(Succeed())
+				Expect(DeployKubeAPIServer(ctx, runtimeClient, namespace, kubeAPIServer, serviceAccountConfig, serverCertificateConfig, sniConfig, externalHostname, nodeNetworkCIDRs, serviceNetworkCIDRs, podNetworkCIDRs, seedPodNetwork, nil, nil, etcdEncryptionKeyRotationPhase, wantScaleDown)).To(Succeed())
 			})
 		})
 
@@ -1557,11 +1566,12 @@ authorizers:
 				kubeAPIServer.EXPECT().SetNodeNetworkCIDRs(gomock.Any())
 				kubeAPIServer.EXPECT().SetServiceNetworkCIDRs(gomock.Any())
 				kubeAPIServer.EXPECT().SetPodNetworkCIDRs(gomock.Any())
+				kubeAPIServer.EXPECT().SetSeedPodNetwork(gomock.Any())
 				kubeAPIServer.EXPECT().SetServerCertificateConfig(gomock.Any())
 				kubeAPIServer.EXPECT().SetServiceAccountConfig(serviceAccountConfig)
 				kubeAPIServer.EXPECT().Deploy(ctx)
 
-				Expect(DeployKubeAPIServer(ctx, runtimeClient, namespace, kubeAPIServer, serviceAccountConfig, serverCertificateConfig, sniConfig, externalHostname, nodeNetworkCIDRs, serviceNetworkCIDRs, podNetworkCIDRs, nil, nil, etcdEncryptionKeyRotationPhase, wantScaleDown)).To(Succeed())
+				Expect(DeployKubeAPIServer(ctx, runtimeClient, namespace, kubeAPIServer, serviceAccountConfig, serverCertificateConfig, sniConfig, externalHostname, nodeNetworkCIDRs, serviceNetworkCIDRs, podNetworkCIDRs, seedPodNetwork, nil, nil, etcdEncryptionKeyRotationPhase, wantScaleDown)).To(Succeed())
 			})
 		})
 
@@ -1575,11 +1585,12 @@ authorizers:
 				kubeAPIServer.EXPECT().SetNodeNetworkCIDRs(gomock.Any())
 				kubeAPIServer.EXPECT().SetServiceNetworkCIDRs(gomock.Any())
 				kubeAPIServer.EXPECT().SetPodNetworkCIDRs(gomock.Any())
+				kubeAPIServer.EXPECT().SetSeedPodNetwork(gomock.Any())
 				kubeAPIServer.EXPECT().SetServerCertificateConfig(gomock.Any())
 				kubeAPIServer.EXPECT().SetServiceAccountConfig(gomock.Any())
 				kubeAPIServer.EXPECT().Deploy(ctx)
 
-				Expect(DeployKubeAPIServer(ctx, runtimeClient, namespace, kubeAPIServer, serviceAccountConfig, serverCertificateConfig, sniConfig, externalHostname, nodeNetworkCIDRs, serviceNetworkCIDRs, podNetworkCIDRs, nil, nil, etcdEncryptionKeyRotationPhase, wantScaleDown)).To(Succeed())
+				Expect(DeployKubeAPIServer(ctx, runtimeClient, namespace, kubeAPIServer, serviceAccountConfig, serverCertificateConfig, sniConfig, externalHostname, nodeNetworkCIDRs, serviceNetworkCIDRs, podNetworkCIDRs, seedPodNetwork, nil, nil, etcdEncryptionKeyRotationPhase, wantScaleDown)).To(Succeed())
 			})
 		})
 
@@ -1593,11 +1604,12 @@ authorizers:
 				kubeAPIServer.EXPECT().SetNodeNetworkCIDRs(gomock.Any())
 				kubeAPIServer.EXPECT().SetServiceNetworkCIDRs(gomock.Any())
 				kubeAPIServer.EXPECT().SetPodNetworkCIDRs(gomock.Any())
+				kubeAPIServer.EXPECT().SetSeedPodNetwork(gomock.Any())
 				kubeAPIServer.EXPECT().SetServerCertificateConfig(gomock.Any())
 				kubeAPIServer.EXPECT().SetServiceAccountConfig(gomock.Any())
 				kubeAPIServer.EXPECT().Deploy(ctx)
 
-				Expect(DeployKubeAPIServer(ctx, runtimeClient, namespace, kubeAPIServer, serviceAccountConfig, serverCertificateConfig, sniConfig, externalHostname, nodeNetworkCIDRs, serviceNetworkCIDRs, podNetworkCIDRs, nil, nil, etcdEncryptionKeyRotationPhase, wantScaleDown)).To(Succeed())
+				Expect(DeployKubeAPIServer(ctx, runtimeClient, namespace, kubeAPIServer, serviceAccountConfig, serverCertificateConfig, sniConfig, externalHostname, nodeNetworkCIDRs, serviceNetworkCIDRs, podNetworkCIDRs, seedPodNetwork, nil, nil, etcdEncryptionKeyRotationPhase, wantScaleDown)).To(Succeed())
 			})
 		})
 	})
