+++ skaffold-gardenadm.yaml
@@ -143,8 +143,9 @@ build:
             - pkg/component/networking/nginxingress
             - pkg/component/networking/nodelocaldns
             - pkg/component/networking/nodelocaldns/constants
+            - pkg/component/networking/vpn/envoy
+            - pkg/component/networking/vpn/envoy/templates/envoy.yaml.tpl
             - pkg/component/networking/vpn/seedserver
-            - pkg/component/networking/vpn/seedserver/templates/envoy.yaml.tpl
             - pkg/component/networking/vpn/shoot
             - pkg/component/nodemanagement/dependencywatchdog
             - pkg/component/nodemanagement/machinecontrollermanager
