+++ skaffold-operator.yaml
@@ -130,8 +130,9 @@ build:
             - pkg/component/networking/istio/charts/istio/istio-istiod
             - pkg/component/networking/nginxingress
             - pkg/component/networking/nodelocaldns/constants
+            - pkg/component/networking/vpn/envoy
+            - pkg/component/networking/vpn/envoy/templates/envoy.yaml.tpl
             - pkg/component/networking/vpn/seedserver
-            - pkg/component/networking/vpn/seedserver/templates/envoy.yaml.tpl
             - pkg/component/observability/logging
             - pkg/component/observability/logging/fluentbit
             - pkg/component/observability/logging/fluentcustomresources
